requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.7
Django==6.0.4
django-allauth==65.15.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-ninja==1.6.2
django-ninja-extra==0.31.4
django-ninja-jwt==5.4.4
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
idna==3.11
injector==0.24.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
